src/main/java/wooteco/subway/domain/Sections.java
package wooteco.subway.domain;

import java.util.HashSet;
import java.util.List;
import java.util.Set;

public class Sections {

    private List<Section> sections;

    public Sections(List<Section> sections) {
        this.sections = sections;
    }

    public Set<Long> extractStationIds() {
        Set<Long> stationIds = new HashSet<>();
        for (Section section : sections) {
            stationIds.add(section.getUpStationId());
            stationIds.add(section.getDownStationId());
        }
        return stationIds;
    }
}
